pyproject.toml
[tool.poetry]
name = "medsearch-api"
version = "0.1.0"
description = ""
authors = ["bexb0t <chrysslyn@gmail.com>"]
readme = "README.md"
packages = [{ include = "medsearch_api",  from="src" }]


[tool.poetry.dependencies]
python = "^3.11"
flask = "^2.3.3"
flask-sqlalchemy = "^3.1.1"
pydantic = "^2.5.3"
alembic = "^1.13.1"
pydantic-settings = "^2.1.0"
mysqlclient = "^2.2.1"
requests = "^2.32.3"
types-requests = "^2.32.0.20240602"
pre-commit = "^3.7.1"

[tool.poetry.group.dev.dependencies]
black = "^23.9.1"
flake8 = "^6.1.0"
pytest = "^7.4.2"
pytest-mock = "^3.11.1"
flake8-length = "^0.3.1"
pre-commit = "^3.4.0"
types-requests = "^2.32.0.20240602"
mypy = "^1.10.0"
flask-sqlalchemy-stubs = "^0.2"
kubernetes = "^30.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
